- NSSpack splitting into one dir for each pb + add/update doxygen doc for Numerics
git-svn-id: svn+ssh://gforge/svn/siconos/trunk@946 bf01a9d7-0df3-0310-8ad4-b1c94d9ae2d5

+++ Numerics/src/netlib/odepack/test/funC.cpp
@@ -15,7 +15,7 @@
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
  *
  * Contact: Vincent ACARY vincent.acary@inrialpes.fr
-*/
+ */
 #include <stdio.h>
 #include "blaslapack.h"
 
@@ -37,7 +37,7 @@ extern "C" void F77NAME(f1)(integer *sizeOfX, doublereal *time, doublereal *x, d
   }
   else
   {
-    printf("OscillatorPlugin:vectorField --- Bad size of x. %i -- %i\n", *sizeOfX, sizeOfX);
+    printf("OscillatorPlugin:vectorField --- Bad size of x. %d -- %d\n", (int)*sizeOfX, (int)sizeOfX);
   }
 
 }
